result of sorted volumes is wrong
The attribute order of mods:part hast to be a positive integer.
As I can see the belonging database field is
tx_dlf_documents.volume_sorting - but it is of the type tinytext.
As a consequence the result of sorted volumes is wrong like this:
[...]/390/40/400/450/50/500/[...]
http://resolver.sub.uni-hamburg.de/goobi/PPN637205685
So I use "CAST ... AS UNSIGNED" to get a correct volume list but it
would be better to change the data type of the affected database field.