toolbox/pdf: Handle special cases for PDF page download in double page view
The old code handled the cases where either left and right page had a
PDF or where there was no PDF for both pages.

The new code also handles the cases where only one of both pages has a PDF.
It then shows the correct download link (either for the left side or for
the right side).

Signed-off-by: Stefan Weil <sw@weilnetz.de>