OAI plugin: Improve error handling for missing configuration
An OAI request with verb=Identity silently failed when configuration data
was missing.

Change this and show default values for admin e-mail and repository name.
Reduce also the severity of the logging from "error" to "notice".

The earliest datestamp does not depend on the configuration.
It is now calculated even when the configuration is incomplete.

The new code does not pass the unneeded character encoding parameter to
htmlspecialchars.

Signed-off-by: Stefan Weil <sw@weilnetz.de>