plugins: Fix some errors in generated HTML code
&lt, &gt must be processed by htmlspecialchars(), otherwise '<', '>'
will be written to the generated HTML code.

Add TRUE as 3rd parameter to all calls of pi_getLL in function
getPageBrowser, so all of those calls use htmlspecialchars().

Example without patch (taken from SUB Hamburg):

This text is shown on the web page:
< - 1 - 2 - 3 - ... - 167 - 168 - 169 - >

This is the first part of the corresponding (buggy) HTML code:
<p class="tx-dlf-pagebrowser_kl">< - 1 - <a href=

Signed-off-by: Stefan Weil <sw@weilnetz.de>