set proper Content-Type if "curlUse" is not set
The implementation of GeneralUtility::getUrl() differ in behaviour if
['TYPO3_CONF_VARS']['SYS']['curlUse'] is set or not.

With curlUse enabled, the full header of all redirects is returned.
Without curlUse only the first header is returned.

In case of a redirection, the Content-Type of the first header is e.g.
"text/html" instead of the desired "application/xml". That's why we
cannot use the fetched Content-Type in this case.

To be save with both implementations, the content-type is calculated
manually with finfo_buffer().