add image download tool
This patch adds a new download tool for the toolbox plugin.

As sub-tools have no flexform configuration, the configuration of the
allowed fileGrps for download is put to toolbox flexform. As done in the
pageview plugin, the best found fileGrp will be used.

Two links will be generated in double mode with the translated labels
"Download left page (JPG)" and "Download right page (JPG)". In single
view mode the label is "Download single page (JPG)". The mime-type is
detected and appended to the label in case of JPG and TIFF.