Update installation instructions for Apache Solr
Also renaming of conf/context.xml to conf/solr.xml.

The name of the file is relevant to access the installation in Tomcat. It might be tricky since the Solr has it’s own solr.xml for configuration purposes, but the name of the file is relevant for Tomcat to generate the path in the URL to access the installation.

Was: context.xml -> localhost:8080/context
Now: solr.xml -> localhost:8080/solr

That wasn’t clear in the old instruction.