Refactor duplicate code
The code for generating output for ListRecords, ListIdentifiers and the resume function was basically the same and was already treated the same in the resume() function itself. I extracted the code into a separate method to reduce the duplication.

The same applies to the generation of the resumption token. It was duplicated in several places and I extracted it into its own method.

In the process I changed some variable names to make clear what they are used for.