Use proper COUNT statement in SQL
The current solution in master does not work because the MYSQL $result
is empty after fetchAll(). You cannot execute a loop with
$result->fetch() afterwards.

The PDOStatement::rowCount() would fix this issue at least for MySQL and
MariaDB but the documentation warns to use it for SELECT queries.

The suggested solution is to make a proper COUNT statement. See the
following issue and patch as example:

https://forge.typo3.org/issues/78129