Replace list() construct with new queryBuilder
The list() function works only reliable with numerical arrays. The
result of the queryBuilder queries are assoziative arrays. Additionally
the order of list() has changed in PHP 7. And the order may change in
the future again.

Best is to avoid list() at all [1].

With PHP 7.2 the current solution doesn't work. The list components are
all NULL.

[1] https://www.php.net/manual/en/function.list.php