Add a new tool "search in document".
This adds a input field to the toolbox in case a fulltext file is
present. The search is executed only in this document. 20 results are
listed and you may page through all results in pages with 20 items.

The search is inspired by the SRU search in DFG-Viewer.

Please note: there is no design provided for this feature (as usual in
Kitodo.Presentation).