Always show volumes of a collection.
In case volumes and parents have different collections assigned it may
happen that these volumes are not shown at all.

E.g.
Parent: "Collection-1"
Volume (child): "Collection-2"

The collection plugin is configured to show only "Collection-2". The
current code cannot find the parent as it does not belogn to "Collection
2". As result, the volume is not shown.

The proposed solution handles the found volume as "toplevel" list entry.
The parent is still not shown.