Fix indexing METS files with linked URIs.
Introduced with #401 it is possible for author metadata fields to link
to the given valueURI attribute.

Unfortunately this feature is not complete in case you are using the
author field for facets. The Solr query breaks due to the not allowed
character 31. We have to split the string in Indexer::processPhysical()
as well.

We should fix this issue in 2.x, too.