Register Plugins without addPItoST43().
The old ExtensionManagementUtility::addPItoST43() is still working in
TYPO3 9 but does not fill the userFunc setting properly.

E.g. for the PageView plugin, the userFunc must contain the real class
namei 'Kitodo\Dlf\Plugin\PageView'. Not the "old" class name tx_dlf_pageview.

The plugin registration is done by TypoScript. This is a way, other core
plugins are doing it as well (e.g. EXT:about).