Show flash messages in indexer using the default queue.
Currently, no feedback is shown on updating records in the TYPO3 list
module. This is due to a the flash message rendering in the backend
module. This use the fluid FlashMessagesViewHelper() which get's passed
the default queuei ('core.template.flashMessages').

I didn't find a way to overwrite this queue. In case of Indexer::add(),
this patch adds the default queue. Messages are shown. But with the
default Bootstrap render including the htmlspecialchars() handling.