Enhance the calendar plugin.
The calendar plugin is not usable without a TypoScrip condition or a
condition inside Fluid. The main() method is without any function so
far.

This patch adds the possibility to place the plugin on a separate page
like the pageview plugin.

The main method now does the type-check. Additionally it is possible to
configure an initial document. E.g. the anchor file of a newspaper. So
you may do a page with a calendar of a special newspaper or a single
year.