Change target navigation of calendar plugin.
The calendar plugin knows the targetPid setting as almost all plugins.
Up to know there is only one targetPid so the pageview has to be on the
same uid as the calendar plugin.

With this patch, the targetPid is only used to address the pageview of
an issue. To link between anchor and year (calendar), the current uid is
used.