Fill orderlabel for sub-entries in TableOfContents plugin.
With newspaper/ephemera year files the fields title, year, volume,
volume_sorting are not filled. The TableOfContents plugin only shows the
structure type ("Year") on the anchor level.

The label we want to see is the mets_orderlabel in this case. But this
field is not queried in the plugin.

Filling the internal "orderlabel" field in the menuArray fixes this
issue. The "orderlabel" will be filled into the "title" field in
getMenuEntry().

Another issue is fixed in this patch: The ordering of the menu entries.
The sorting does not work if volume is not filled. In this case the
sorting by mets_orderlabel must be used.