Fix JavaScript error introduced with PR 487.
Using the rotate buttons, a JavaScript error is shown in the webconsole:

``Uncaught TypeError: Cannot read property 'rotate' of undefined``

This is due to the change of the condition in OL3.js:

Old:
```
if (this.ov_view != null) {
	this.ov_view.rotate(0);
}
```

New in PR 487:
```
if (this.ov_view !== null) {
	this.ov_view.rotate(0);
}
```

This patch:
```
if (this.ov_view !== null && this.ov_view !== undefined) {
	this.ov_view.rotate(0);
}
```

The problem is, that `this.ov_view` is always `undefined` NOT `null`. So
the change with PR #487 causes the entry into the if condition.

I suspect that the above condition will never work. But I have no idea
how the "magnifier" feature works and didn't get it working.