Fix value attri of search input field
If you do a phrase search with double quoates, the search input field
does not contain the search phrase after return.

Example: Search for "Leipzig Plagwitz".
<pre>
	<label for="tx-dlf-search-query">&quot;Leipzig Plagwitz&quot;</label>
	<input type="text" id="tx-dlf-search-query" name="tx_dlf[query]" value=""Leipzig Plagwitz"" />
</pre>
Whis is interpreted by the browser as:
<pre>
	<label for="tx-dlf-search-query">"Leipzig Plagwitz"</label>
	<input type="text" id="tx-dlf-search-query" name="tx_dlf[query]" value="" leipzig="" plagwitz""="">
</pre>

In case of the `<label>`-tag everything is ok. In case of the
`<input>`-tag, the HTML-markup is broken.