Fix search in documents for hierarchical documents like newspaper.
== Current Situation

The search plugin has the option to restrict the search in current
document and/or in current collection. This can be used to place the
plugin on a separate page to search only in one collection. Or you place
the plugin on the workview page and trigger a search in document.

Of course, the SearchInDocument plugin is better suited for the workview as the results are reported by via AJAX and the visitor will stay on the current page.

On the opposite, using the search plugin makes it possible to search
in all issues of a newspaper year or even a full newspaper title. At
least, this should be possible. The current implementation does not work
as expected.

=== Example

==== Search plugin on workview (issue)

For example, you can place the plugin on level of a newspaper issue. If
you launch a search, the current implementation will search not only in
current issue but in all issues of the current year. This is a
frequently demanded feature but works only by hasard.

newspaper title (anchor)
   |
   |-> year
        |
        |-> Issue

The current search plugin does two mistakes:

1. Instead of the current document uid, the parent uid is set.
2. The search is done in given uid and all children.

==== Search plugin on workview (year/anchor)

If you place the search plugin on the workview with calendar view, the
search fails.

Same procedure as above. The search will be done in the anchor and it's
children (years). All these files have no fulltexts. So the result is
always empty.

== Proposed Implementation

The proposed implementation does change the searchIn document limitation
of the search plugin to search in current document _and_ all it's children.

This is done on using Solr join feature. To make this possible, the
partof field must use the docValues feature as "uid" does it already.