Fix for SearchInDocument plugin if using routeEnhancers (Slug)
With TYPO3 9.5 the advanced page routing is builtin. There is no need to
install any third party extension like realurl anymore.

Using the routeEnhancer with the pageview makes it possible to write
shorter URLs like

  http://example.com/workview/123/1

which is short for

  http://example.com/workview?tx_dlf[id]=123&tx_dlf[page]=1

The current implementation of the SearchInDocument plugin has to build
the target URLs from the searchresults in JavaScript. It does it by
appending the GET-parameter to the base_url. The resulting URL could
look like this:

  http://example.com/workview/123/1?tx_dlf[id]=123&tx_dlf[page]=5&tx_dlf[highlight_word=searchword

In fact the GET parameter are given twice to TYPO3. This was tolerated
in the past by realurl on TYPO3 8.7 but brings an exception with TYPO3
9.5.

The way the plugin calculates the base URL in JavaScript is even to
complicate. It would be sufficiant to take it from the form action URL.
This way, the double GET parameters are avoided.