Migrate DocumentType-Swith to Symfonie Expression Language.
TYPO3 9.4 introduced the Symfonie Expression Language in the core [1].
This is used for TypoScript conditions, too and the new standard as of
TYPO3 9.5. The old TypoScript conditions are marked as deprecated.

This patch adds the new DocumentTypeProvider to enable the new syntax.
The old user function was kept as it still works and both switches may
co-exist and the old switch is still needed in TYPO3 8.7.

An example usage has been documented.

[1] https://docs.typo3.org/c/typo3/cms-core/master/en-us/Changelog/9.4/Feature-85828-MoveSymfonyExpressionLanguageHandlingIntoEXTcore.html
[2] https://docs.typo3.org/c/typo3/cms-core/master/en-us/Changelog/9.5/Deprecation-86068-OldConditionSyntax.html?highlight=typoscript%20conditions