Proposal: Show uptime in format like "uptime" on linux console.
uptime on Linux returns:

  23:41:18 up 4 days,  6:37,  1 user,  load average: 0,91, 0,99, 1,10

The first is the current time, followed by "up", the amount of days
which might be more than a year, followed by h:m.

The proposed output for Solr-uptime would be now:

  Uptime: 0 days, 00:22:56

This includes the seconds, too.

The solution before was:

  Uptime: 0:21:59.385

Which I find harder to read and not self-explaining.