Use mets_orderlabel for sorting as fallback.
This happens in case of newspaper (year).

The current workaround of using title_sorting is not working for
documents following the METS application profile as the title and
title_sorting are empty for structtype year.

Only METS ORDERLABEL is set, which is used in the calendar view. But it
is not used in listview. That's why only one year is shown below a
newspaper title (anchor) in single collection view.