Check for extension configuration before access.
From TYPO3 10.4 a method protected method `ExtensionConfiguration::hasConfiguration()` is introduced to check the availability of the extnension configuration before access. Anyway, an excep
tion is thrown if the configuration is not present.

Bis patch will check like the `hasConfiguration()` method before `get()`.

The deprecation note #82254 tells us that the array
$GLOBALS['TYPO3_CONF_VARS']['EXT']['extConf'] is deprecated and has been
_replaced_ by $GLOBALS['TYPO3_CONF_VARS']['EXTENSIONS']. So there is no
need to access this array after `ExtensionConfiguration::get()` as we
have already done it.