Avoid logger warning if using the documentType switch
The following warning is logged for structure types newspaper, ephemera and year.

   [WARNING] request="5206c9c7bf5ef" component="Kitodo.Dlf.Common.MetsDocument": Invalid PID 0 for metadata definitions

This is caused, because the given cPid is not transfered to the
`MetsDocument::getMetadata()` function. It get's lost in
`Document::getTitleData` on calling `addMetadataFromMets()`.

This patch avoids the warning. The result of the condition keeps the
same. But I'm not sure, if this is the clean way.