Search with HTTP POST may fail with TYPO3 9.5.
This is always the case, if the target page is cached by mistake. This
can easily happen, if a search engine or anyone else calls the target
page with some GET parameter but of course without a valid
cHash-parameter.

In this case, TYPO3 9.5 safes the page to cache including the wrong
parameters.

A search sending the parameters by POST will receive the cached page
though the parameter 'tx_dlf[id]' is passed which should disable
caching.

This is a know bug (comparable with [1] and can be worked around by
sending the search parameters by HTTP GET.

[1] https://forge.typo3.org/issues/88631