Allow new solr records even without Solr connected.
On creating a new tx_dlf_solrcores record, the corresponding core is
created on the connected Solr automatically. This is nice for the first
setup.

If you change the Solr server, move Kitodo.Presentation to a new server
or do some fancy stuff like using an external readonly Solr for search
only, you have to manipulate the tx_dlf_solrcores records 'by hand'.

This patch gives the full freedom to TYPO3 admins to create records
independantly to an established Solr connection.

As long, as there is no connection, a red banner informs the user. If
the connection is etablished later and the corresponding core is
present, the core stats is shown. This is no new feature.