Allow setting of Solr core name on creating solrcore records.
We allowed to edit the index_name field recently. It stayed a required
field but anything you typed in was ignored and the next possible
index_name was set (e.g. "dlfCore1").

With this change, it is possible to create cores with the given name.
Something like "mySolrcore" is possible now.

Additionally, the field is not "required" anymore. So it may be empty.
In this case, the automatic behavior is kept. A index_name is set and a
Solr core created.