Fix TYPO3@9 compat w.r.t. Extbase repositories
In TYPO3@10, when calling `makeInstance` to instantiate an Extbase
repository, the object manager is automatically injected. This is not
yet the case in TYPO3@9.

This uses the Inject annotation where possible, and goes via
`ObjectManager::get` where not. We will need to change this once we
support TYPO3@11 and drop support for TYPO3@9.