This adds the core mapping of sys_file_reference again.
The documentTypeSwitch has a strange behavior with TYPO3v9.
Sometimes the class mapping array is empty. That's why we set
the class mapping with Helper::polyfillExtbaseClassesForTYPO3v9().

Even worse is the missing of the mapping for the sys_file_reference
table. This is required by tx_dlf_collections.thumbnail field.

This patch adds this mapping to the extension class mapping. This should
be removed if support for TYPO3v9 is dropped.