Tweak resumptionToken tag. (#1)
* Always output the resumptionToken tag.

This adds again the resumptionToken if we have a complete list without a
token itself. This is not a must feature but quite common and I think
it's not forbidden.

This changes the XSLT stylesheet a bit because we output the "cursor" in
the bottom which is not the amount of submitted records but the position
in the list of records.

Just another extample of a complete list:

https://www.tib.eu/oai/public/repository/open?verb=ListIdentifiers&metadataPrefix=oai_dc&set=zbmql

* Revert "Always output the resumptionToken tag."

Convinced. I should return an empty resumptionToken tag even if it is
not forbidden.

This reverts commit 079265aa1a187da119c439cd70ea705bc3458ae5.

* Change missleading label